test(display-runtime): move the scene-clock tests to frame_ingest_test.rs
`render_thread` keeps unit tests in a sibling `<name>_test.rs` declared
with `#[path]` (`cursor_test.rs`, `state_test.rs`, `child_frames_test.rs`);
the scene-clock tests were an inline module in `frame_ingest.rs` that the
placement-cache commit extended in place.  Move them where the crate
keeps the others, unchanged.
